src/test/java/com/transfer/transferMoney/service/UserServiceImplTest.java
package com.transfer.transferMoney.service;

import com.transfer.transferMoney.Repository.UserRepository;
import com.transfer.transferMoney.model.User;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.ArgumentCaptor;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;

import static org.assertj.core.api.Assertions.*;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;
@ExtendWith(MockitoExtension.class)
class UserServiceImplTest {
    @Mock
    private UserRepository userRepository;
    private UserServiceImpl underTest;

    @BeforeEach
    void setUp() {
        underTest = new UserServiceImpl(userRepository);

    }



    @Test
    void canSearchByDni() {
        // GIVEN
        Long dni = 12345678L;
        User user = new User();
        user.setDni(dni);
        //WHEN
        underTest.searchByDni(dni);
        //THEN
        verify(userRepository).findByDni(dni);

    }
    @Test
    void canAddUser(){
        //GIVE
        User user = User.builder()
                .email("example@gmail.com")
                .password("example")
                .dni(12345678L)
                .username("exampleUser")
                .build();
        //WHEN
        underTest.save(user);

        //THEN
        ArgumentCaptor<User> userArgumentCaptor = ArgumentCaptor.forClass(User.class);
        verify(userRepository).save(userArgumentCaptor.capture());
        User capturedUser = userArgumentCaptor.getValue();
        assertThat(capturedUser).isEqualTo(user);
    }
}